Telegram.Bot: update to 16.0.2
Mostly to handle long user identifiers.

+++ CaptchaBot/Services/ChatUser.cs
@@ -4,9 +4,9 @@
     {
         public long ChatId { get; }
 
-        public int UserId { get; }
+        public long UserId { get; }
 
-        public ChatUser(long chatId, int userId)
+        public ChatUser(long chatId, long userId)
         {
             ChatId = chatId;
             UserId = userId;
